+++ src/django/api/tests/test_production_locations_partial_update.py
@@ -286,20 +286,43 @@ class TestProductionLocationsPartialUpdate(APITestCase):
         expected_response_body = {
             'detail': 'The request body is invalid.',
             'errors': [
-                {'field': 'sector',
-                 'detail': ('Field sector must be a string or a'
-                            ' list of strings.')},
-                {'field': 'location_type',
-                 'detail': ('Field location_type must be a string or a list of'
-                            ' strings.')},
-                {'field': 'number_of_workers',
-                 'errors': [
-                     {'field': 'min',
-                      'detail': ('Ensure this value is greater than or equal'
-                                 ' to 1.')},
-                     {'field': 'max',
-                      'detail': ('Ensure this value is greater than or equal'
-                                 ' to 1.')}]}]}
+                {
+                    'field': 'sector',
+                    'detail':
+                        (
+                            'Field sector must be '
+                            'a string or a list of strings.'
+                        )
+                },
+                {
+                    'field': 'location_type',
+                    'detail': (
+                        'Field location_type must be a string'
+                        ' or a list of strings.'
+                    )
+                },
+                {
+                    'field': 'number_of_workers',
+                    'errors': [
+                        {
+                            'field': 'min',
+                            'detail': (
+                                'Ensure this value is greater than'
+                                ' or equal to 1.'
+                            )
+                        },
+                        {
+                            'field': 'max',
+                            'detail': (
+                                'Ensure this value is greater than'
+                                ' or equal to 1.'
+                            )
+                        }
+                    ]
+                }
+            ]
+        }
+
         initial_moderation_event_count = ModerationEvent.objects.count()
 
         invalid_req_body = json.dumps({
@@ -307,12 +330,16 @@ class TestProductionLocationsPartialUpdate(APITestCase):
             'name': 'Blue Horizon Facility',
             'address': '990 Spring Garden St., Philadelphia PA 19123',
             'country': 'US',
-            'sector': {'some_key': 1135},
+            'sector': {
+                'some_key': 1135
+            },
             'parent_company': 'string',
             'product_type': [
                 'string'
             ],
-            'location_type': {'some_key': 1135},
+            'location_type': {
+                'some_key': 1135
+            },
             'processing_type': [
                 'string'
             ],
@@ -339,3 +366,107 @@ class TestProductionLocationsPartialUpdate(APITestCase):
         # Ensure that no ModerationEvent record has been created.
         self.assertEqual(ModerationEvent.objects.count(),
                          initial_moderation_event_count)
+
+    @patch('api.geocoding.requests.get')
+    def test_moderation_event_not_created_with_invalid_parent_company(
+            self,
+            mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = geocoding_data
+
+        expected_response_body = {
+            'detail': 'The request body is invalid.',
+            'errors': [
+                {
+                    'field': 'parent_company',
+                    'detail': (
+                        'Field parent_company must be a string'
+                        ' not a number.'
+                    )
+                },
+            ]
+         }
+        initial_moderation_event_count = ModerationEvent.objects.count()
+
+        invalid_req_body = json.dumps({
+            'source': 'API',
+            'name': 'Blue Horizon Facility',
+            'address': '990 Spring Garden St., Philadelphia PA 19123',
+            'country': 'US',
+            'parent_company': 12345,
+        })
+
+        response = self.client.patch(
+            self.url,
+            invalid_req_body,
+            content_type='application/json'
+        )
+        self.assertEqual(response.status_code,
+                         status.HTTP_422_UNPROCESSABLE_ENTITY)
+
+        response_body_dict = json.loads(response.content)
+        self.assertEqual(response_body_dict, expected_response_body)
+        self.assertEqual(ModerationEvent.objects.count(),
+                         initial_moderation_event_count)
+
+    @patch('api.geocoding.requests.get')
+    def test_moderation_event_created_with_valid_parent_company(
+            self,
+            mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = geocoding_data
+
+        special_characters = '&@, \' _ #()'
+        numbers = '1234567890'
+        multi_lang_letters = '贾建龙ÖrmeTİCіїъыParentCompanyการผลิตהפָקָהผลิต'
+        valid_parent_company = (
+            special_characters +
+            numbers +
+            multi_lang_letters
+        )
+
+        valid_req_body = json.dumps({
+            'source': 'SLC',
+            'name': 'Blue Horizon Facility',
+            'address': '990 Spring Garden St., Philadelphia PA 19123',
+            'country': 'US',
+            'parent_company': valid_parent_company
+        })
+
+        response = self.client.patch(
+            self.url,
+            valid_req_body,
+            content_type='application/json'
+        )
+        self.assertEqual(response.status_code, 202)
+
+        response_body_dict = json.loads(response.content)
+        response_moderation_id = response_body_dict.get('moderation_id')
+        moderation_event = ModerationEvent.objects.get(
+            pk=response_moderation_id
+        )
+        stringified_created_at = moderation_event.created_at.strftime(
+            '%Y-%m-%dT%H:%M:%S.%f'
+        ) + 'Z'
+
+        self.assertEqual(
+            response_body_dict.get('moderation_status'),
+            'PENDING'
+        )
+        self.assertEqual(
+            response_body_dict.get('created_at'),
+            stringified_created_at
+        )
+        self.assertEqual(
+            response_moderation_id,
+            str(moderation_event.uuid)
+        )
+        self.assertIn("cleaned_data", response_body_dict)
+        parent_company = (
+            response_body_dict
+            .get('cleaned_data', {})
+            .get('fields', {})
+            .get('parent_company')
+        )
+        self.assertEqual(len(response_body_dict), 4)
+        self.assertEqual(parent_company, valid_parent_company)